% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_tdwg_list.R
\name{build_tdwg_list}
\alias{build_tdwg_list}
\title{Get raw data from TDWG}
\usage{
build_tdwg_list()
}
\value{
list of tibbles
}
\description{
Rough function to populate a list of data from TDWG, with (all?) information
needed to look up terms, versions etc.
}
\details{
Later on, it might be useful to 1. cache this information somewhere, and
2. add checks for whether information is up to date
}
