% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{collect.tdwg}
\alias{collect.tdwg}
\title{complete a TDWG pipe}
\usage{
\method{collect}{tdwg}(x, ...)
}
\arguments{
\item{.query}{An object of class \code{tdwg}, created with \code{tdwg()}}
}
\description{
This function evaluates a \code{tdwg} object
}
