% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwc_terms.R
\name{dwc_terms}
\alias{dwc_terms}
\title{Get DwC terms that match user-specified arguments}
\usage{
dwc_terms(date = "2023-09-18")
}
\arguments{
\item{date}{(string) A valid date on which a DwC standard was released.}
}
\value{
A \code{tibble} listing the requested terms, along with appropriate
metadata.
}
\description{
Basically a wrapper function to the piped syntax given by \code{tdwg()} et al.
}
