% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_terms.R
\name{get_terms}
\alias{get_terms}
\title{get DwC terms and format them}
\usage{
get_terms()
}
\description{
This was originally in \code{galaxias}, but is a dependency of \code{view_terms()}, so
is included here. This content probably needs more thought, and to be
integrated with other functions.
Also, functions like this should return tibbles, not lists, unless as part of
a named class (i.e. \code{tdwg}).
}
