% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{show_}
\alias{show_}
\alias{show_standards}
\alias{show_vocabularies}
\alias{show_termlists}
\alias{show_terms}
\title{Show TDWG data}
\format{
All \code{tibble}s returned by \code{show_} functions have the following columns:
\describe{
\item{code}{Identifier for the type of data within the column; i.e.
for \code{show_standards()} this is the unique code for that
standard.}
\item{date}{Release date for selected object in \code{"YYYY-MM-DD"} format.
For standards, this is effectively synonymous with a version of
the standard; but this is not the case at lower levels of the
hierarchy.}
\item{label}{Human-readable tag for this datum.}
\item{description}{A detailed overview of the function of this datum.}
\item{status}{\code{"recommended"}, \code{"superseded"} or \code{"deprecated"}.}
\item{key}{A unique identifier used to navigate the information hierarchy,
formed as a URL.}
}

In addition, \code{terms} data have the following additional columns:
\describe{
\item{type}{TDWG terms contain descriptions of the terms themselves
(type = \code{terms}), but also the classes that are related to
those terms (type = \code{class}). Note that terms can be nested
within classes or vice versa.}
\item{parent_class}{If a \code{term} is nested within a \code{class}, the parent
class is listed here. Useful for grouping like terms.}
}
}
\usage{
show_standards()

show_vocabularies()

show_termlists()

show_terms()
}
\description{
Low-level functions for calling internal data sourced from TDWG. These
functions are not full exports of TDWG data structures; some columns have
been removed, and others have been renamed. For piping, use \code{tdwg_} functions
instead.
}
