% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tawnydragon-packge.R
\docType{package}
\name{tawnydragon}
\alias{tawnydragon}
\alias{tawnydragon-package}
\title{Biodiversity Data Standards from TDWG}
\description{
This package provides information on biodiversity standards built and
maintained by TDWG. It is experimental, designed to provide quick, easy
access to information needed to build, validate or interpret Darwin Core
archives. All of the data contained here is built by TDWG and is provided
unmodified.
}
\section{Functions}{

\strong{subfunctions}
\itemize{
\item \code{\link[=get_standards]{get_standards()}}
\item \code{\link[=get_vocabularies]{get_vocabularies()}}
\item \code{\link[=get_termlists]{get_termlists()}}
\item \code{\link[=get_terms]{get_terms()}}
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/AtlasOfLivingAustralia/tawnydragon}
  \item Report bugs at \url{https://github.com/AtlasOfLivingAustralia/tawnydragon/issues}
}

}
\author{
\strong{Maintainer}: Martin Westgate \email{martin.westgate@csiro.au}

}
\keyword{internal}
