% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdwg.R
\name{tdwg}
\alias{tdwg}
\alias{print.tdwg}
\title{Open a pipe for TDWG data}
\usage{
tdwg(version = "recommended")

\method{print}{tdwg}(x, ...)
}
\description{
This creates an object of class \code{tdwg}, which can be piped to refine the
information returned at the end. This is useful for navigating down the
TDWG information hierarchy
}
