% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdwg_standards.R, R/tdwg_vocabularies.R,
%   R/tdwg_termlists.R
\name{tdwg_standards}
\alias{tdwg_standards}
\alias{tdwg_vocabularies}
\alias{tdwg_termlists}
\title{Amend a \code{tdwg} query}
\usage{
tdwg_standards(.query, standard = "Darwin Core", version = "recommended")

tdwg_vocabularies(.query, vocabulary = NULL, version = NULL)

tdwg_termlists(.query, version = NULL, termlist = NULL)
}
\arguments{
\item{.query}{An object of class \code{tdwg}, created with \code{tdwg()}}

\item{standard}{(string) Either a valid \code{label} (or part thereof, e.g.
\code{"Audibon"} will return \code{"Audibon Core"}), or a TDWG standard number (e.g.
\code{"450"}). See \code{\link[=valid_standards]{valid_standards()}} for details.
#' @param version (string) Either \code{"recommended"} (the default), or a valid
date string in \code{"YYYY-MM-DD"} format. See \code{\link[=valid_versions]{valid_versions()}} for examples.
Only used if not already specified by an earlier function.}

\item{vacabulary}{(string) Either a valid \code{label} (or part thereof, e.g.
\code{"Audibon"} will return \code{"Audibon Core"}), or a TDWG vocabulary code (e.g.
\code{"dwc"}). See \code{\link[=valid_vacabularies]{valid_vacabularies()}} for details.}
}
\description{
These functions drill down through the TDWG information hierarchy
}
\details{
The sequence of \code{tdwg} objects is \code{standards} > \code{vocabularies} > \verb{term lists}
(> \code{classes}) > \code{terms}.
}
