% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdwg_classes.R
\name{tdwg_classes}
\alias{tdwg_classes}
\title{Get classes}
\usage{
tdwg_classes(.x, ...)
}
\description{
\code{class}es are a type of \code{term}. They are nested within \code{termLists} within the
TDWG hierarchy.
}
\details{
Many terms are within classes; for example class \code{Event} includes an
\code{eventID}, \code{eventTime} and many others. Ergo classes are useful for grouping
similar types of \code{class}. However, some \code{class}es are nested within terms;
\code{LivingSpecimen} is best thought of as a valid entry to the \code{basisOfRecord}
\code{term}, for example.
}
