% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terms.R
\name{terms.tdwg}
\alias{terms.tdwg}
\title{Get terms from a TDWG pipe}
\usage{
\method{terms}{tdwg}(x, ...)
}
\description{
Equivalent to \code{collect.tdwg}, but returns a \code{vector} rather than \code{tibble}.
}
