% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid.R
\name{valid_standards}
\alias{valid_standards}
\alias{valid_versions}
\alias{valid_vocabularies}
\title{Look up which standards are supported}
\usage{
valid_standards()

valid_versions(standard)

valid_vocabularies()
}
\arguments{
\item{standard}{(optional) if given, a standard for which valid versions will
be given. Otherwise, all dates with releases will be shown}
}
\description{
Look up which standards are supported
}
