% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_terms.R
\name{view_terms}
\alias{view_terms}
\title{View DwC terms in the browser}
\usage{
view_terms()
}
\description{
This is a prototype initially imported from \code{galaxias}. Requires more care
to optimize for \code{tawnydragon}.
}
