% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_all.R
\name{vocabularies}
\alias{vocabularies}
\title{look up vocabularies}
\usage{
vocabularies(.x, standard, version, ...)
}
\arguments{
\item{.x}{a single-row \code{tibble} giving the required standard, created using
\code{standard()}}

\item{standard}{(optional) string giving the standard for which to display
valid vocabularies. Not required if \code{.x} is given}

\item{version}{(optional) string giving the version for which to display
valid vocabularies. Not required if \code{.x} is given}

\item{...}{arguments passed to \code{dplyr::filter}}
}
\description{
look up vocabularies
}
