% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdwg.R
\name{tdwg}
\alias{tdwg}
\alias{tdwg_standards}
\alias{tdwg_vocabularies}
\alias{tdwg_termlists}
\alias{tdwg_terms}
\alias{summarize.tdwg}
\alias{summarise.tdwg}
\title{Piped syntax for navigating TDWG data}
\usage{
tdwg()

tdwg_standards(.query, ...)

tdwg_vocabularies(.query, ...)

tdwg_termlists(.query, ...)

tdwg_terms(.query, ...)

\method{summarize}{tdwg}(.data, ...)

\method{summarise}{tdwg}(.data, ...)
}
\arguments{
\item{.query}{An object of class \code{tdwg}, created with \code{tdwg()}}

\item{...}{Further arguments passed to \code{dplyr::filter()}}
}
\value{
An object of class \code{tdwg}, which is a length-4 list containing
the requested information; except \code{summarize()} which returns a \code{tibble}
}
\description{
This creates an object of class \code{tdwg}, which can be piped to refine the
information returned at the end. Susequent functions amend the subsequent
query. To investigate complete tibbles without piping, use \code{show_} functions.
}
\details{
An unusual feature of this syntax is that requests propagate up the hierarchy,
but not down. That is, calling \code{tdwg_standards()} (level 1) fills only the
\code{standards} slot, but calls to \code{tdwg_vocabularies()} (level 2) fills
\code{vocabularies} and \code{standards}. The interpretation of 'back-filled' slots
is that they show all entries that contain at least one row from the selected
level of the hierarchy. It does \emph{not} answer the question "Which single
standard contains all the terms I asked for?".
}
