% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_all.R
\name{vocabularies}
\alias{vocabularies}
\title{look up vocabularies}
\usage{
vocabularies(.x, ...)
}
\arguments{
\item{.x}{a single-row \code{tibble} giving the required standard, created using
\code{standard()}}

\item{...}{arguments passed to \code{dplyr::filter}}
}
\description{
look up vocabularies
}
